Add `--all-sheets` option
The all-sheets option now allows me to replace instances of code that
looked like:

```
for sheet in (excelchop -p w file.xlsx)
     excelchop -w $sheet file.xlsx
end
```

with

```
excelchop --all-sheets file.xlsx
```

minimizing the number of calls to the executable.

This also now allows the sheets specified in the `-w` option to be able
to be split on newlines. This is probably the better case where the list
of sheets is dynamically being built up using some normal Unix filter
like awk.